Refactor: Rename root package and remove onboarding feature
- Renamed the `decompose` package to `root` for better clarity.
- Removed the `Onboarding` feature, including its related files and references in `Root.kt`, `RootComponent.kt`, and `RootConfig.kt`.
