+++ composeApp/src/commonMain/kotlin/ru/bortsov/holdmaster/composeapp/root/Root.kt
@@ -1,14 +1,13 @@
-package ru.bortsov.holdmaster.composeapp.decompose
+package ru.bortsov.holdmaster.composeapp.root
 
 import com.arkivanov.decompose.ComponentContext
 import com.arkivanov.decompose.router.slot.ChildSlot
 import com.arkivanov.decompose.router.stack.ChildStack
 import com.arkivanov.decompose.value.Value
 import com.arkivanov.essenty.backhandler.BackHandlerOwner
-import ru.bortsov.holdmaster.composeapp.decompose.feature.onboarding.Onboarding
-import ru.bortsov.holdmaster.composeapp.decompose.feature.tabs.Tabs
-import ru.bortsov.holdmaster.composeapp.decompose.splash.Splash
 import ru.bortsov.holdmaster.composeapp.error.ErrorDialog
+import ru.bortsov.holdmaster.composeapp.splash.Splash
+import ru.bortsov.holdmaster.composeapp.tabs.Tabs
 import ru.bortsov.holdmaster.feature.auth.presentation.navigation.Auth
 import ru.bortsov.holdmaster.feature.photo.presentation.component.Photo
 
@@ -25,7 +24,6 @@ interface Root: BackHandlerOwner {
         class SplashChild(val component: Splash) : Child()
         class TabsChild(val component: Tabs) : Child()
         class AuthChild(val component: Auth) : Child()
-        class OnboardingChild(val component: Onboarding) : Child()
         class TakePhotoChild(val component: Photo) : Child()
     }
