+++ composeApp/src/commonMain/kotlin/ru/bortsov/holdmaster/composeapp/root/RootComponent.kt
@@ -1,4 +1,4 @@
-package ru.bortsov.holdmaster.composeapp.decompose
+package ru.bortsov.holdmaster.composeapp.root
 
 import com.arkivanov.decompose.ComponentContext
 import com.arkivanov.decompose.router.slot.ChildSlot
@@ -13,10 +13,9 @@ import com.arkivanov.decompose.router.stack.pop
 import com.arkivanov.decompose.router.stack.popTo
 import com.arkivanov.decompose.router.stack.replaceAll
 import com.arkivanov.decompose.value.Value
-import ru.bortsov.holdmaster.composeapp.decompose.feature.onboarding.Onboarding
-import ru.bortsov.holdmaster.composeapp.decompose.feature.tabs.Tabs
-import ru.bortsov.holdmaster.composeapp.decompose.splash.Splash
 import ru.bortsov.holdmaster.composeapp.error.ErrorDialog
+import ru.bortsov.holdmaster.composeapp.splash.Splash
+import ru.bortsov.holdmaster.composeapp.tabs.Tabs
 import ru.bortsov.holdmaster.feature.auth.presentation.navigation.Auth
 import ru.bortsov.holdmaster.feature.photo.presentation.component.Photo
 
@@ -25,7 +24,6 @@ class RootComponent(
     private val splashComponentFactory: Splash.Factory,
     private val authComponentFactory: Auth.Factory,
     private val tabsComponentFactory: Tabs.Factory,
-    private val onboardingComponentFactory: Onboarding.Factory,
     private val errorDialogFactory: ErrorDialog.Factory,
     private val takePhotoFactory: Photo.Factory,
 ) : Root, ComponentContext by componentContext {
@@ -81,11 +79,6 @@ class RootComponent(
             )
         )
 
-
-        RootConfig.Stack.Onboarding -> {
-            Root.Child.OnboardingChild(onboardingComponentFactory(componentContext))
-        }
-
         RootConfig.Stack.Tabs -> {
             Root.Child.TabsChild(tabsComponentFactory(componentContext))
         }
@@ -112,7 +105,6 @@ class RootComponent(
         private val splashComponentFactory: Splash.Factory,
         private val authComponentFactory: Auth.Factory,
         private val tabsComponentFactory: Tabs.Factory,
-        private val onboardingComponentFactory: Onboarding.Factory,
         private val errorDialogFactory: ErrorDialog.Factory,
         private val takePhotoFactory: Photo.Factory,
     ) : Root.Factory {
@@ -122,7 +114,6 @@ class RootComponent(
                 componentContext = componentContext,
                 authComponentFactory = authComponentFactory,
                 tabsComponentFactory = tabsComponentFactory,
-                onboardingComponentFactory = onboardingComponentFactory,
                 errorDialogFactory = errorDialogFactory,
                 takePhotoFactory = takePhotoFactory,
             )
